# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HRUO Enterprise HCM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recruitment/jobs:
    get:
      operationId: listJobs
      tags: [recruitment]
      summary: List job postings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [recruitment]
      summary: Create job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /recruitment/jobs/{id}:
    patch:
      operationId: updateJob
      tags: [recruitment]
      summary: Update job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /recruitment/applicants:
    get:
      operationId: listApplicants
      tags: [recruitment]
      summary: List applicants
      parameters:
        - name: jobId
          in: query
          schema:
            type: ["integer", "null"]
        - name: stage
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Applicant"
    post:
      operationId: createApplicant
      tags: [recruitment]
      summary: Add applicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
  /recruitment/applicants/{id}:
    patch:
      operationId: updateApplicant
      tags: [recruitment]
      summary: Update applicant stage/status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicantUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"

  # ── EXPENSES ──────────────────────────────────────────────────────────────
...
  /performance/goals:
    get:
      operationId: listGoals
      tags: [performance]
      summary: List performance goals
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [performance]
      summary: Create goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /performance/goals/{id}:
    patch:
      operationId: updateGoal
      tags: [performance]
      summary: Update goal progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /performance/reviews:
    get:
      operationId: listReviews
      tags: [performance]
      summary: List performance reviews
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [performance]
      summary: Create performance review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ── TRAINING ──────────────────────────────────────────────────────────────
  /training/courses:
    get:
      operationId: listCourses
      tags: [training]
      summary: List training courses
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [training]
      summary: Create training course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /training/records:
    get:
      operationId: listTrainingRecords
      tags: [training]
      summary: List employee training records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingRecord"
    post:
      operationId: createTrainingRecord
      tags: [training]
      summary: Add training record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecord"

  # ── DOCUMENTS ─────────────────────────────────────────────────────────────
...
  /quotes:
    get:
      operationId: listQuoteRequests
      tags: [quotes]
      summary: List quote requests
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteRequest"
    post:
      operationId: createQuoteRequest
      tags: [quotes]
      summary: Submit quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
  /quotes/{id}:
    patch:
      operationId: updateQuoteRequest
      tags: [quotes]
      summary: Update quote request (super admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"

  # ── SUPER ADMIN ───────────────────────────────────────────────────────────
  /superadmin/tenants:
    get:
      operationId: listTenants
      tags: [superadmin]
      summary: List all tenants (super admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [superadmin]
      summary: Create tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /superadmin/tenants/{id}:
    patch:
      operationId: updateTenant
      tags: [superadmin]
      summary: Update tenant status / subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
...
  /superadmin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [superadmin]
      summary: List audit logs
      parameters:
        - name: tenantId
          in: query
          schema:
            type: ["integer", "null"]
        - name: action
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, status]
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published, closed]
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        applicantCount:
          type: integer
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
        description:
          type: string
        requirements:
          type: string
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        description:
          type: string
        requirements:
          type: string
    Applicant:
      type: object
      required: [id, name, email, jobId, stage]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        jobId:
          type: integer
        jobTitle:
          type: ["string", "null"]
        stage:
          type: string
          enum: [applied, screening, interview, technical, offer, hired, rejected]
        rating:
          type: ["number", "null"]
        resumeUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        appliedAt:
          type: string
    ApplicantInput:
      type: object
      required: [name, email, jobId]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        jobId:
          type: integer
        resumeUrl:
          type: string
    ApplicantUpdate:
      type: object
      properties:
        stage:
          type: string
        rating:
          type: number
        notes:
          type: string

    # ── EXPENSES ──────────────────────────────────────────────────────────────
...
    Goal:
      type: object
      required: [id, title, employeeId, progress, status]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        targetDate:
          type: ["string", "null"]
        weight:
          type: number
        progress:
          type: number
        status:
          type: string
          enum: [not_started, in_progress, completed, cancelled]
        type:
          type: string
          enum: [company, department, individual]
    GoalInput:
      type: object
      required: [title, employeeId]
      properties:
        title:
          type: string
        description:
          type: string
        employeeId:
          type: integer
        targetDate:
          type: string
        weight:
          type: number
        type:
          type: string
    GoalUpdate:
      type: object
      properties:
        progress:
          type: number
        status:
          type: string
        description:
          type: string
    Review:
      type: object
      required: [id, employeeId, period, rating, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        reviewerId:
          type: ["integer", "null"]
        reviewerName:
          type: ["string", "null"]
        period:
          type: string
        rating:
          type: number
        strengths:
          type: ["string", "null"]
        improvements:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, submitted, acknowledged]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [employeeId, period, rating]
      properties:
        employeeId:
          type: integer
        period:
          type: string
        rating:
          type: number
        strengths:
          type: string
        improvements:
          type: string

    # ── TRAINING ──────────────────────────────────────────────────────────────
    Course:
      type: object
      required: [id, name, code, active]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        active:
          type: boolean
        location:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        trainerName:
          type: ["string", "null"]
    CourseInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        location:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        description:
          type: string
    TrainingRecord:
      type: object
      required: [id, employeeId, courseId, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [enrolled, in_progress, completed, cancelled]
        rating:
          type: ["number", "null"]
        feedback:
          type: ["string", "null"]
    TrainingRecordInput:
      type: object
      required: [employeeId, courseId]
      properties:
        employeeId:
          type: integer
        courseId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string

    # ── DOCUMENTS ─────────────────────────────────────────────────────────────
...
    QuoteRequest:
      type: object
      required: [id, companyName, contactName, email, status]
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        modules:
          type: array
          items:
            type: string
        employeeCount:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, quoted, converted, closed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    QuoteRequestInput:
      type: object
      required: [companyName, contactName, email]
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        modules:
          type: array
          items:
            type: string
        employeeCount:
          type: integer
        message:
          type: string
    QuoteRequestUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string

    # ── APPROVAL ──────────────────────────────────────────────────────────────
...
    Tenant:
      type: object
      required: [id, name, slug, status]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [pending, active, suspended, trial]
        logo:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        employeeCount:
          type: integer
        subscriptionStart:
          type: ["string", "null"]
        subscriptionEnd:
          type: ["string", "null"]
        modules:
          type: array
          items:
            type: string
        createdAt:
          type: string
    TenantInput:
      type: object
      required: [name, slug, contactEmail]
      properties:
        name:
          type: string
        slug:
          type: string
        contactEmail:
          type: string
        modules:
          type: array
          items:
            type: string
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        logo:
          type: string
        subscriptionStart:
          type: string
        subscriptionEnd:
          type: string
        modules:
          type: array
          items:
            type: string
        notes:
          type: string

    # ── SUPER ADMIN ANALYTICS ─────────────────────────────────────────────────
...
    AuditLog:
      type: object
      required: [id, action, actorEmail, timestamp]
      properties:
        id:
          type: integer
        action:
          type: string
        actorEmail:
          type: string
        tenantId:
          type: ["integer", "null"]
        tenantName:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        timestamp:
          type: string